Environments

JuliaRL.AbstractEnvironment — Type.

Represents an abstract environment for reinforcement learning agents. Has several functions that need to be implemented to work

source

JuliaRL.start! — Method.

start!(env::AbstractEnvironment; rng::AbstractRNG, kwargs...)

Function to start the passed environment env.

source

JuliaRL.start — Method.

start(env::AbstractEnvironment; rng::AbstractRNG, kwargs...)

Function to retrieve a started environment struct without effecting the current environment struct. Returns an environment of the same type.

source

JuliaRL.step! — Method.

step!(env::AbstractEnvironment, action; rng::AbstractRNG, kwargs...)

Update the state of the passed environment env based on the underlying dynamics and the action.

source

JuliaRL.step — Method.

step!(env::AbstractEnvironment, action; rng::AbstractRNG, kwargs...)

Update the state of the environment based on the underlying dynamics and the action.

source

The above functions take advantage of the following interface.

JuliaRL.reset! — Method.

reset!(env::AbstractEnvironment; rng::AbstractRNG, kwargs...)

Reset the environment to initial conditions based on the random number generator.

source

JuliaRL.environment_step! — Method.

environment_step!(env::AbstractEnvironment, action; rng::AbstractRNG, kwargs...)

Update the state of the environment based on the underlying dynamics and the action. This is not used directly, but through the step function.

source

JuliaRL.get_reward — Method.

get_reward(env::AbstractEnvironment)

Retrieve reward for the current state of the environment.

source

JuliaRL.is_terminal — Method.

is_terminal(env::AbstractEnvironment)

Check if the environment is in a terminal state

source

JuliaRL.get_state — Method.

get_state(env::AbstractEnvironment)

Retrieve the current state of the environment

source

JuliaRL.get_actions — Method.

get_actions(env::AbstractEnvironment)

Returns the set of actions available to take.

source

To dig into the dynamics of the environment while running, we can also implement two interfaces for visualization.

JuliaRL.Base.show(io::IO, env::AbstractEnvironment)
JuliaRL.render(env::AbstractEnvironment, args...; kwargs...)