Learning

Learning

Abstract API

JuliaRL.Learning.AbstractValueFunction — Type.

AbstractValueFunctionn

Abstract type definition for a value function

source

JuliaRL.Learning.AbstractVFunction — Type.

AbstractVFunction

Abstract type definition for a state value function

source

JuliaRL.Learning.AbstractQFunction — Type.

AbstractQFunction

Abstract type definition for a state-action value function

source

JuliaRL.Learning.update! — Function.

update!(value::ValueFunction, opt::Optimizer, ρ, s_t, s_tp1, reward, γ, terminal)

Arguments:

value::ValueFunction: opt::Optimizer: ρ: Importance sampling ratios (Array of Floats) s_t: States at time t s_tp1: States at time t + 1 reward: cumulant or reward for value function γ: discount factor terminal: Determining termination of the episode (if applicable).

source

update!(value::ValueFunction, opt::Optimizer, ρ, s_t, s_tp1, reward, γ, terminal)

Arguments:

value::ValueFunction: opt::Optimizer: ρ: Importance sampling ratios (Array of Floats) s_t: States at time t s_tp1: States at time t + 1 reward: cumulant or reward for value function γ: discount factor terminal: Determining termination of the episode (if applicable). a_t: Action at time t a_tp1: Action at time t + 1 target_policy: Action at time t

source

LinearRL

JuliaRL.Learning.LinearRL.VFunction — Type.

VFunction(num_features)

A structure hosting the weights for a linear value function. Used for Linear function approximation.

source

JuliaRL.Learning.LinearRL.SparseVFunction — Type.

SparseVFunction(num_features)

A structure for when the feature vector is known to be sparse, and handed to the agent as a list of indices. Significantly faster than a normal value function in this special case.

source

JuliaRL.Learning.LinearRL.QFunction — Type.

QFunction(num_features, num_features_per_action, num_actions)

Linear QFunction. Assumes no sparsity, and an array of floats for a feature vector.

source

JuliaRL.Learning.LinearRL.SparseQFunction — Type.

SparseQFunction(num_features, num_features_per_action, num_actions)

QFunction assuming sparsity.

source

Currently implemented algorithms

JuliaRL.Learning.LinearRL.TD — Type.

TD(α)

Online Temporal Difference Learning.

source

JuliaRL.Learning.LinearRL.TDLambda — Type.

TDLambda(α, λ)

Online Temporal Difference Learning with eligibility traces.

source

JuliaRL.Learning.LinearRL.WatkinsQ — Type.

WatkinsQ(α)

Q Learning as defined by Watkins

source